model.RecordKey -> schema.RecordKey
