+++ internal/pkg/service/buffer/store/recordstore/record.go
@@ -7,19 +7,19 @@ import (
 
 	"github.com/c2h5oh/datasize"
 
-	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/model"
+	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/model/schema"
 	serviceError "github.com/keboola/keboola-as-code/internal/pkg/service/common/errors"
 	"github.com/keboola/keboola-as-code/internal/pkg/telemetry"
 	"github.com/keboola/keboola-as-code/internal/pkg/utils/errors"
 )
 
-func (c *Store) CreateRecord(ctx context.Context, recordKey model.RecordKey, csvData []string) (err error) {
+func (c *Store) CreateRecord(ctx context.Context, recordKey schema.RecordKey, csvData []string) (err error) {
 	tracer, client := c.tracer, c.etcdClient
 
 	_, span := tracer.Start(ctx, "keboola.go.buffer.recordstore.CreateRecord")
 	defer telemetry.EndSpan(span, &err)
 
-	key := recordKey.Key()
+	key := recordKey.In(c.schema)
 	csvBuffer := new(bytes.Buffer)
 	w := csv.NewWriter(csvBuffer)
